requirements2.txt
asgiref==3.8.1
attrs==25.1.0
autobahn==24.4.2
Automat==24.8.1
black==25.1.0
cffi==1.17.1
channels==4.2.0
click==8.1.8
constantly==23.10.4
cryptography==44.0.2
daphne==4.1.2
Django==5.1.6
django-phonenumber-field==8.1.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
isort==6.0.1
loguru==0.7.3
lxml==5.3.1
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
phonenumbers==9.0.3
platformdirs==4.3.6
plum-dispatch==1.7.4
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pyOpenSSL==25.0.0
python-docx==1.1.2
python-dotenv==1.0.1
service-identity==24.2.0
setuptools==75.8.2
Spire.Doc==13.1.0
sqlparse==0.5.3
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
zope.interface==7.2
